Refactor tests to mirror main package structure.
